Sort entries (#15)
This adds a preference to sort entries by name, or keep the previous behavior which is keep the entries unsorted.

The default behavior is now to sort the entries by name.

It felt awkward to put it anywhere but into a general settings fragment, which did not exist yet, so this PR also adds a tiny settings fragment that can be accessed for the action menu of the ConnectionsFragment.

⚠ Because I added the French translation (and attempted to provide good enough German translation, sorry if it's not!), this PR requires validation of PR #14 first, thus the WIP status. Otherwise it should be fine!

Let me know if you wish some things to be done differently!

Closes #13

Co-authored-by: dece <shgck@pistache.land>
Reviewed-on: https://codeberg.org/qwerty287/ftpclient/pulls/15
Co-authored-by: dece <dece@noreply.codeberg.org>
Co-committed-by: dece <dece@noreply.codeberg.org>